Move header parser code to HeaderKeys
