+++ core/src/main/scala/org/http4s/headers/Content-Disposition.scala
@@ -2,11 +2,18 @@ package org.http4s
 package headers
 
 import org.http4s.Header.Raw
+import org.http4s.parser.Http4sHeaderParser
 import org.http4s.util.Writer
 
 object `Content-Disposition` extends HeaderKey.Internal[`Content-Disposition`] with HeaderKey.Singleton {
-  override protected def parseHeader(raw: Raw): Option[`Content-Disposition`.HeaderT] =
-    parser.SimpleHeaders.CONTENT_DISPOSITION(raw.value).toOption
+  override protected def parseHeader(raw: Raw): Option[`Content-Disposition`] = {
+    new Http4sHeaderParser[`Content-Disposition`](raw.value) {
+      def entry = rule {
+        Token ~ zeroOrMore(";" ~ OptWS ~ Parameter) ~ EOL ~> { (token:String, params: Seq[(String, String)]) =>
+          `Content-Disposition`(token, params.toMap)}
+      }
+    }.parse.toOption
+  }
 }
 
 // see http://tools.ietf.org/html/rfc2183
