+++ core/src/main/scala/org/http4s/headers/Content-Length.scala
@@ -2,11 +2,15 @@ package org.http4s
 package headers
 
 import org.http4s.Header.Raw
+import org.http4s.parser.Http4sHeaderParser
 import org.http4s.util.Writer
 
 object `Content-Length` extends HeaderKey.Internal[`Content-Length`] with HeaderKey.Singleton  {
-  override protected def parseHeader(raw: Raw): Option[`Content-Length`.HeaderT] =
-    parser.SimpleHeaders.CONTENT_LENGTH(raw.value).toOption
+  override protected def parseHeader(raw: Raw): Option[`Content-Length`] = {
+    new Http4sHeaderParser[`Content-Length`](raw.value) {
+      def entry = rule { Digits ~ EOL ~> {s: String => `Content-Length`(s.toInt)} }
+    }.parse.toOption
+  }
 }
 
 final case class `Content-Length`(length: Int) extends Header.Parsed {
