+++ core/src/main/scala/org/http4s/headers/Cookie.scala
@@ -1,14 +1,22 @@
-package org.http4s
-package headers
+package org.http4s.headers
 
 import org.http4s.Header.Raw
+import org.http4s.{Header, HeaderKey}
+import org.http4s.parser.BaseCookieParser
 import org.http4s.util.Writer
+import org.parboiled2._
 
 import scalaz.NonEmptyList
 
 object Cookie extends HeaderKey.Internal[Cookie] with HeaderKey.Recurring {
-  override protected def parseHeader(raw: Raw): Option[Cookie.HeaderT] =
-    parser.CookieHeader.COOKIE(raw.value).toOption
+  override protected def parseHeader(raw: Raw): Option[Cookie] =
+    new CookieParser(raw.value).parse.toOption
+
+  private class CookieParser(input: ParserInput) extends BaseCookieParser[Cookie](input) {
+    def entry: Rule1[Cookie] = rule {
+      oneOrMore(CookiePair).separatedBy(";" ~ OptWS) ~ EOI ~> {xs: Seq[org.http4s.Cookie] => Cookie(xs.head, xs.tail: _*)}
+    }
+  }
 }
 
 final case class Cookie(values: NonEmptyList[org.http4s.Cookie]) extends Header.RecurringRenderable {
