+++ core/src/main/scala/org/http4s/headers/ETag.scala
@@ -2,11 +2,17 @@ package org.http4s
 package headers
 
 import org.http4s.Header.Raw
+import org.http4s.parser.Http4sHeaderParser
 import org.http4s.util.Writer
 
 object ETag extends HeaderKey.Internal[ETag] with HeaderKey.Singleton {
-  override protected def parseHeader(raw: Raw): Option[ETag.HeaderT] =
-    parser.SimpleHeaders.ETAG(raw.value).toOption
+  override protected def parseHeader(raw: Raw): Option[ETag] = {
+    new Http4sHeaderParser[ETag](raw.value) {
+      def entry = rule {
+        capture(zeroOrMore(AlphaNum)) ~> (ETag(_))
+      }
+    }.parse.toOption
+  }
 }
 
 final case class ETag(tag: String) extends Header.Parsed {
