+++ core/src/main/scala/org/http4s/headers/If-Modified-Since.scala
@@ -2,11 +2,17 @@ package org.http4s
 package headers
 
 import org.http4s.Header.Raw
+import org.http4s.parser.Http4sHeaderParser
 import org.http4s.util.Writer
 
 object `If-Modified-Since` extends HeaderKey.Internal[`If-Modified-Since`] with HeaderKey.Singleton {
-  override protected def parseHeader(raw: Raw): Option[`If-Modified-Since`] =
-    parser.SimpleHeaders.IF_MODIFIED_SINCE(raw.value).toOption
+  override protected def parseHeader(raw: Raw): Option[`If-Modified-Since`] = {
+    new Http4sHeaderParser[`If-Modified-Since`](raw.value) {
+      def entry = rule {
+        HttpDate ~ EOL ~> (`If-Modified-Since`(_))
+      }
+    }.parse.toOption
+  }
 }
 
 final case class `If-Modified-Since`(date: DateTime) extends Header.Parsed {
