+++ core/src/main/scala/org/http4s/headers/If-None-Match.scala
@@ -2,11 +2,17 @@ package org.http4s
 package headers
 
 import org.http4s.Header.Raw
+import org.http4s.parser.Http4sHeaderParser
 import org.http4s.util.Writer
 
 object `If-None-Match` extends HeaderKey.Internal[`If-None-Match`] with HeaderKey.Singleton {
-  override protected def parseHeader(raw: Raw): Option[`If-None-Match`.HeaderT] =
-    parser.SimpleHeaders.IF_NONE_MATCH(raw.value).toOption
+  override protected def parseHeader(raw: Raw): Option[`If-None-Match`] = {
+    new Http4sHeaderParser[`If-None-Match`](raw.value) {
+      def entry = rule {
+        capture(zeroOrMore(AlphaNum)) ~> (`If-None-Match`(_))
+      }
+    }.parse.toOption
+  }
 }
 
 case class `If-None-Match`(tag: String) extends Header.Parsed {
