+++ core/src/main/scala/org/http4s/headers/Range.scala
@@ -2,6 +2,7 @@ package org.http4s
 package headers
 
 import org.http4s.Header.Raw
+import org.http4s.parser.{RangeRule, Http4sHeaderParser}
 import org.http4s.util.{Renderable, Writer}
 
 import scalaz.NonEmptyList
@@ -9,10 +10,6 @@ import scalaz.NonEmptyList
 // See https://tools.ietf.org/html/rfc7233
 
 object Range extends HeaderKey.Internal[Range] with HeaderKey.Singleton {
-
-
-  override protected def parseHeader(raw: Raw): Option[Range.HeaderT] =
-    parser.RangeParser.RANGE(raw.value).toOption
 
   def apply(unit: RangeUnit, r1: SubRange, rs: SubRange*): Range =
     Range(unit, NonEmptyList(r1, rs:_*))
@@ -37,7 +34,15 @@ object Range extends HeaderKey.Internal[Range] with HeaderKey.Singleton {
     }
   }
 
-
+  override protected def parseHeader(raw: Raw): Option[Range] = {
+    new Http4sHeaderParser[Range](raw.value) with RangeRule {
+      def entry = rule {
+        capture(oneOrMore(Alpha)) ~ '=' ~ oneOrMore(byteRange).separatedBy(',') ~> { (s: String, rs: Seq[SubRange]) =>
+          Range(RangeUnit(s), NonEmptyList(rs.head, rs.tail:_*))
+        }
+      }
+    }.parse.toOption
+  }
 }
 
 case class Range(unit: RangeUnit, ranges: NonEmptyList[Range.SubRange]) extends Header.Parsed {
