+++ core/src/main/scala/org/http4s/headers/X-Forwarded-For.scala
@@ -4,13 +4,22 @@ package headers
 import java.net.InetAddress
 
 import org.http4s.Header.Raw
+import org.http4s.parser.Http4sHeaderParser
 import org.http4s.util.Writer
 
 import scalaz.NonEmptyList
 
 object `X-Forwarded-For` extends HeaderKey.Internal[`X-Forwarded-For`] with HeaderKey.Recurring {
-  override protected def parseHeader(raw: Raw): Option[`X-Forwarded-For`.HeaderT] =
-    parser.SimpleHeaders.X_FORWARDED_FOR(raw.value).toOption
+  override protected def parseHeader(raw: Raw): Option[`X-Forwarded-For`] = {
+    new Http4sHeaderParser[`X-Forwarded-For`](raw.value) {
+      def entry = rule {
+        oneOrMore((Ip ~> (Some(_)))  | ("unknown" ~ push(None))).separatedBy(ListSep) ~
+          EOL ~> { xs: Seq[Option[InetAddress]] =>
+          `X-Forwarded-For`(xs.head, xs.tail: _*)
+        }
+      }
+    }.parse.toOption
+  }
 }
 
 final case class `X-Forwarded-For`(values: NonEmptyList[Option[InetAddress]]) extends Header.Recurring {
