+++ core/src/main/scala/org/http4s/parser/ChallengeRules.scala
@@ -3,7 +3,8 @@ package org.http4s.parser
 import org.http4s.{Header, Challenge}
 import org.parboiled2._
 
-abstract private[parser] class ChallengeParser[H <: Header](input: ParserInput) extends Http4sHeaderParser[H](input) {
+/** Base methods used by `Proxy-Authenticate` and `WWW-Authenticate`. */
+abstract private[http4s] class ChallengeRules[H <: Header](input: ParserInput) extends Http4sHeaderParser[H](input) {
 
   def ChallengeRule: Rule1[Challenge] = rule {
     Token ~ oneOrMore(LWS) ~ zeroOrMore(AuthParam).separatedBy(ListSep) ~> {
