+++ core/src/main/scala/org/http4s/parser/Http4sHeaderParser.scala
@@ -5,7 +5,7 @@ import org.parboiled2._
 import scalaz.{\/, Validation}
 
 
-private[parser] abstract class Http4sHeaderParser[H <: Header](val input: ParserInput) extends Parser with AdditionalRules  {
+private[http4s] abstract class Http4sHeaderParser[H <: Header](val input: ParserInput) extends Parser with AdditionalRules  {
 
   def entry: Rule1[H]
